requirements.txt
# Core dependencies for AI Requests System

# Database
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Embeddings
sentence-transformers>=2.2.2

# LLM & Fine-tuning
transformers>=4.35.0
torch>=2.1.0
peft>=0.7.0
accelerate>=0.24.0
bitsandbytes>=0.41.0  # For 4-bit quantization (required for RAG)
huggingface_hub>=0.19.0  # For downloading models

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Utilities
tqdm>=4.66.0

# SQL Server (if needed for export)
pyodbc>=5.0.0  # Optional - only if exporting from SQL Server

